\name{illuminaNormalisation-package}
\alias{illuminaNormalisation-package}
\alias{illuminaNormalisation}
\docType{package}
\title{
Scripts for running background correction, normalization and quality assessment of Illumina Beadchip data.
}
\description{
Scripts for running background correction, normalization and quality assessment of Illumina Beadchip data using the methods found in the limma and lumi packages on the openCPU platform provided by TNO. If possible, missing packages are installed.
}
\details{
\tabular{ll}{
Package: \tab illuminaNormalisation\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-03-17\cr
License: \tab Apache 2.0\cr
}
Do not run locally, use openCPU to direct this package as it is a pipeline.
}
\author{
Job van Riet (Original scripts provided by: Varshna Goelela + ArrayAnalysis.org)

Maintainer: Job van Riet <job.vanriet@tno.nl>
}
\references{
No :>
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\examples{
Do not run locally, use openCPU to direct this package as it is a pipeline.
}
